Extrated Field() into its own class
